update :: Long -> Integer, stockId -> stockCode 변경

+++ src/main/java/com/juicycool/backend/domain/stock/repository/StockRepository.java
@@ -3,5 +3,5 @@ package com.juicycool.backend.domain.stock.repository;
 import com.juicycool.backend.domain.stock.Stock;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StockRepository extends JpaRepository<Stock, Long> {
+public interface StockRepository extends JpaRepository<Stock, Integer> {
 }
